Add solution for "Most Common Word" problem and refactor ProblemLink property
Introduced a new solution for the LeetCode problem "Most Common Word". Also, updated the property `ProblemLink` in several problem classes and the `ILeetCodeProblem` interface to no longer have a setter, making it get-only. The "Minimum Operations to Form Subsequence With Target Sum" was also replaced in the main program with the newly implemented problem, "Most Common Word".